Self To System v1.1, added description dialogs

+++ SelfToSystem/app/src/main/java/com/cypher/selftosystem/MainActivity.java
@@ -3,23 +3,36 @@ package com.cypher.selftosystem;
 import android.app.Activity;
 import android.app.AlertDialog;
 import android.content.DialogInterface;
+import android.content.Intent;
+import android.content.SharedPreferences;
+import android.net.Uri;
 import android.os.Bundle;
+import android.os.Handler;
+import android.os.Looper;
+import android.os.Message;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
-import com.cypher.selftosystem.SystemAppUtilities.SystemAppManagementException;
+import com.cypher.selftosystem.SystemAppUtilities.SystemAppUtilitiesException;
 
 public class MainActivity extends Activity {
 	public static final String TAG = "SelfToSystem";
+
+	public static final String PREF = TAG + "_Pref";
+	public static final String DESCRIPTION = TAG + "_Pref";
 
 	private TextView tv_rootAvailable;
 	private TextView tv_userApp;
 	private TextView tv_systemApp;
 
+	private SharedPreferences pref;
+
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.activity_main);
+
+		pref = getSharedPreferences(PREF, MODE_PRIVATE);
 
 		tv_rootAvailable = (TextView) findViewById(R.id.tv_rootAvailable);
 		tv_userApp = (TextView) findViewById(R.id.tv_userApp);
@@ -34,10 +47,12 @@ public class MainActivity extends Activity {
 		btn_refreshData.setOnClickListener(new View.OnClickListener() {
 			@Override
 			public void onClick(View view) {
-				try {
-					refreshData();
-				} catch (SystemAppManagementException e) {
-					errorDialog(e.getMessage());
+				if (blockingOkCancelDialog(getString(R.string.msg_header_info), getString(R.string.msg_refreshData))) {
+					try {
+						refreshData();
+					} catch (SystemAppUtilitiesException e) {
+						errorDialog(e.getMessage());
+					}
 				}
 			}
 		});
@@ -45,10 +60,12 @@ public class MainActivity extends Activity {
 		btn_gainRootAccess.setOnClickListener(new View.OnClickListener() {
 			@Override
 			public void onClick(View view) {
-				try {
-					SystemAppUtilities.gainRootAccess();
-				} catch (SystemAppManagementException e) {
-					errorDialog(e.getMessage());
+				if (blockingOkCancelDialog(getString(R.string.msg_header_info), getString(R.string.msg_gainRootAccess))) {
+					try {
+						SystemAppUtilities.gainRootAccess();
+					} catch (SystemAppUtilitiesException e) {
+						errorDialog(e.getMessage());
+					}
 				}
 			}
 		});
@@ -56,10 +73,12 @@ public class MainActivity extends Activity {
 		btn_installAsSystemApp.setOnClickListener(new View.OnClickListener() {
 			@Override
 			public void onClick(View view) {
-				try {
-					SystemAppUtilities.installAsSystemApp(MainActivity.this, true);
-				} catch (SystemAppManagementException e) {
-					errorDialog(e.getMessage());
+				if (blockingOkCancelDialog(getString(R.string.msg_header_warning), getString(R.string.msg_installAsSystemApp))) {
+					try {
+						SystemAppUtilities.installAsSystemApp(MainActivity.this, true);
+					} catch (SystemAppUtilitiesException e) {
+						errorDialog(e.getMessage());
+					}
 				}
 			}
 		});
@@ -67,10 +86,12 @@ public class MainActivity extends Activity {
 		btn_uninstallSystemApp.setOnClickListener(new View.OnClickListener() {
 			@Override
 			public void onClick(View view) {
-				try {
-					SystemAppUtilities.uninstallSystemApp(MainActivity.this, true);
-				} catch (SystemAppManagementException e) {
-					errorDialog(e.getMessage());
+				if (blockingOkCancelDialog(getString(R.string.msg_header_warning), getString(R.string.msg_uninstallSystemApp))) {
+					try {
+						SystemAppUtilities.uninstallSystemApp(MainActivity.this, true);
+					} catch (SystemAppUtilitiesException e) {
+						errorDialog(e.getMessage());
+					}
 				}
 			}
 		});
@@ -78,10 +99,12 @@ public class MainActivity extends Activity {
 		btn_uninstallUserApp.setOnClickListener(new View.OnClickListener() {
 			@Override
 			public void onClick(View view) {
-				try {
-					SystemAppUtilities.uninstallUserApp(MainActivity.this);
-				} catch (SystemAppManagementException e) {
-					errorDialog(e.getMessage());
+				if (blockingOkCancelDialog(getString(R.string.msg_header_warning), getString(R.string.msg_uninstallUserApp))) {
+					try {
+						SystemAppUtilities.uninstallUserApp(MainActivity.this);
+					} catch (SystemAppUtilitiesException e) {
+						errorDialog(e.getMessage());
+					}
 				}
 			}
 		});
@@ -89,33 +112,49 @@ public class MainActivity extends Activity {
 		btn_clearAppData.setOnClickListener(new View.OnClickListener() {
 			@Override
 			public void onClick(View view) {
-				try {
-					SystemAppUtilities.clearAppData(MainActivity.this);
-				} catch (SystemAppManagementException e) {
-					errorDialog(e.getMessage());
+				if (blockingOkCancelDialog(getString(R.string.msg_header_warning), getString(R.string.msg_clearAppData))) {
+					try {
+						SystemAppUtilities.clearAppData(MainActivity.this);
+					} catch (SystemAppUtilitiesException e) {
+						errorDialog(e.getMessage());
+					}
 				}
 			}
 		});
+
+		if (pref.getBoolean(DESCRIPTION, true)) {
+			pref.edit().putBoolean(DESCRIPTION, false).apply();
+
+			if (!blockingTwoButtonDialog(getString(R.string.app_name), getString(R.string.msg_first_start), getString(R.string.button_close), getString(R.string.button_github))) {
+				Uri uri = Uri.parse("https://github.com/Cypher01/SelfToSystem");
+				Intent intent = new Intent(Intent.ACTION_VIEW, uri);
+				startActivity(intent);
+			}
+		}
 
 		if (!SystemAppUtilities.isRootAvailable()) {
 			btn_gainRootAccess.setEnabled(false);
 			btn_installAsSystemApp.setEnabled(false);
 			btn_uninstallSystemApp.setEnabled(false);
+			btn_uninstallUserApp.setEnabled(false);
+			btn_clearAppData.setEnabled(false);
 
-			errorDialog("This App works only on a rooted device. It won't help without root access.");
+			errorDialog(getString(R.string.msg_root_unavailable));
 		}
 	}
 
 	@Override
 	protected void onResume() {
+		super.onResume();
+
 		try {
 			refreshData();
-		} catch (SystemAppManagementException e) {
+		} catch (SystemAppUtilitiesException e) {
 			e.printStackTrace();
 		}
 	}
 
-	private void refreshData() throws SystemAppManagementException {
+	private void refreshData() throws SystemAppUtilitiesException {
 		if (SystemAppUtilities.isRootAvailable()) {
 			tv_rootAvailable.setText(getString(R.string.root_available));
 			tv_rootAvailable.setTextColor(0xff99cc00); // @android:color/holo_green_light
@@ -128,32 +167,72 @@ public class MainActivity extends Activity {
 		tv_systemApp.setText(SystemAppUtilities.getApkInfos(MainActivity.this, true));
 	}
 
+	public boolean blockingOkCancelDialog(String title, String message) {
+		return blockingTwoButtonDialog(title, message, getString(R.string.button_ok), getString(R.string.button_cancel));
+	}
+
+	public boolean blockingTwoButtonDialog(String title, String message, String button1, String button2) {
+		final boolean[] mResult = new boolean[1]; // for whatever reason, IntelliJ suggested to use an array
+
+		final Handler handler = new Handler() {
+			@Override
+			public void handleMessage(Message mesg) {
+				throw new RuntimeException();
+			}
+		};
+
+		AlertDialog.Builder builder = new AlertDialog.Builder(MainActivity.this);
+		builder.setTitle(title)
+				.setMessage(message)
+				.setPositiveButton(button1, new DialogInterface.OnClickListener() {
+					public void onClick(DialogInterface dialog, int whichButton) {
+						mResult[0] = true;
+						handler.sendMessage(handler.obtainMessage());
+					}
+				})
+						.setNegativeButton(button2, new DialogInterface.OnClickListener() {
+					public void onClick(DialogInterface dialog, int whichButton) {
+						mResult[0] = false;
+						handler.sendMessage(handler.obtainMessage());
+					}
+				})
+				.show();
+
+		try { Looper.loop(); } catch (RuntimeException ignored) { }
+
+		return mResult[0];
+	}
+
+	public boolean blockingOneButtonDialog(String button, String title, String message) {
+		final boolean[] mResult = new boolean[1]; // for whatever reason, IntelliJ suggested to use an array
+
+		final Handler handler = new Handler() {
+			@Override
+			public void handleMessage(Message mesg) {
+				throw new RuntimeException();
+			}
+		};
+
+		AlertDialog.Builder builder = new AlertDialog.Builder(MainActivity.this);
+		builder.setTitle(title)
+				.setMessage(message)
+				.setNeutralButton(button, new DialogInterface.OnClickListener() {
+					public void onClick(DialogInterface dialog, int whichButton) {
+						mResult[0] = true;
+						handler.sendMessage(handler.obtainMessage());
+					}
+				}).show();
+
+		try { Looper.loop(); } catch (RuntimeException ignored) { }
+
+		return mResult[0];
+	}
+
 	private void errorDialog(String message) {
 		AlertDialog.Builder builder = new AlertDialog.Builder(MainActivity.this);
-		builder.setTitle("Error")
+		builder.setTitle(getString(R.string.msg_header_error))
 				.setMessage(message)
 				.setNeutralButton(R.string.button_close, null)
 				.show();
-	}
-
-	private boolean warningDialog(String message) {
-		final boolean[] answer = new boolean[1]; // for whatever reason, IntelliJ suggested to use an array
-		AlertDialog.Builder builder = new AlertDialog.Builder(MainActivity.this);
-		builder.setTitle("Warning")
-				.setMessage(message)
-				.setPositiveButton(R.string.button_ok, new DialogInterface.OnClickListener() {
-					@Override
-					public void onClick(DialogInterface dialog, int which) {
-						answer[0] = true;
-					}
-				})
-				.setNegativeButton(R.string.button_cancel, new DialogInterface.OnClickListener() {
-					@Override
-					public void onClick(DialogInterface dialog, int which) {
-						answer[0] = false;
-					}
-				})
-				.show();
-		return answer[0];
 	}
 }
